feat(lib): add package-level functions
Similar to some packages of the standard libarary (e.g. log) Notify now
has a global but private standard instance that gets created on initialization.
This is more of a convenience feature as it potentially saves a single
step when interacting with Notify. Functions like 'UseServices' and
'Send' are now usable directly from the package.

In the following we assume 's' is a fully initialized Notify service.

Example(old):

  n := notify.New()
  n.UseServices(s)
  n.Send(...)

Example(new):

  notify.UseServices(s)
  notify.Send(...)

Of course the New function still exists! This is a non-breaking change
that, in a sense, just adds some syntactic-sugar.